+++ Digital Pantry/Model/SectionOfPantry.swift
@@ -11,14 +11,14 @@ import Foundation
 
 struct SectionOfPantry {
     //properties/members of Recipe_Diet
-    var sectionID: Int
+    var sectionID: Int64
     var section: String
-    var appPantryID: Int
-    //var appUserID: Int
+    var appPantryID: Int64
+    //var appUserID: Int64
     
     
     //initialised where we have not obtained a UniqueIdentifier (recipe_DietID) as yet
-    init?(section: String, appPantryID: Int) {
+    init?(section: String, appPantryID: Int64) {
         self.sectionID = -1
         self.section = section
         self.appPantryID = appPantryID
@@ -26,17 +26,17 @@ struct SectionOfPantry {
     }
     
     //use case: calling data back from DB/table
-    init?(sectionID: Int, section: String, appPantryID: Int) {
+    init?(sectionID: Int64, section: String, appPantryID: Int64) {
         self.sectionID = sectionID
         self.section = section
         self.appPantryID = appPantryID
     }
     
-    mutating func setSectionID(sectionID: Int) { //once the DB has a sectionID, use this to set it
+    mutating func setSectionID(sectionID: Int64) { //once the DB has a sectionID, use this to set it
         self.sectionID = sectionID
     }
     
-    func getSectionID() -> Int { //retrieve sectionID
+    func getSectionID() -> Int64 { //retrieve sectionID
         return sectionID
     }
         
@@ -48,11 +48,11 @@ struct SectionOfPantry {
         self.section = section
     }
 
-    func getAppPantryID() -> Int { //retrieve appPantryID
+    func getAppPantryID() -> Int64 { //retrieve appPantryID
         return appPantryID
     }
     
-    mutating func setAppPantryID(appPantryID: Int) { //add a appPantryID to item
+    mutating func setAppPantryID(appPantryID: Int64) { //add a appPantryID to item
         self.appPantryID = appPantryID
     }
 }
